+++ src/mes_utils.hpp
@@ -9,6 +9,6 @@ std::vector<std::string> equal_shares_utils(
     const std::vector<std::string>& voters,
     const std::vector<std::string>& projects,
     const std::unordered_map<std::string, double>& cost,
-    const std::unordered_map<std::string, std::vector<std::pair<std::string, int>>>& approvers_utilities,
-    const std::unordered_map<std::string, int>& total_utility,
+    const std::unordered_map<std::string, std::vector<std::pair<std::string, long long>>>& approvers_utilities,
+    const std::unordered_map<std::string, long long>& total_utility,
     double total_budget);
